refactor(electron): Electron file layout, IPC  and errors. * Changed file layout of core and modules on electron side. * Refactored IPC code to be more structured. Created shell-manager. * Added structured error codes and message mapping for toasts.
