+++ .vscode/settings.json
@@ -5,5 +5,6 @@
     "[javascript]": {
         "editor.defaultFormatter": "esbenp.prettier-vscode"
     },
-    "eslint.alwaysShowStatus": true
+    "eslint.alwaysShowStatus": true,
+    "typescript.tsserver.experimental.enableProjectDiagnostics": true,
 }
