+++ electron/core/shell-manager/evaluator/_eval.ts
