+++ electron/main.ts
@@ -1,7 +1,6 @@
-import { app, protocol } from "electron";
+import { app, BrowserWindowConstructorOptions, protocol, BrowserWindow } from "electron";
 import path from "path";
 
-import Window from "./modules/window";
 import IPC from "./modules/ipc";
 
 import { enableDevTools } from "./utils/dev";
@@ -18,27 +17,31 @@ import { ARK_FOLDER_PATH } from "./utils/constants";
 			callback({ path: `${ARK_FOLDER_PATH}${url}` });
 		});
 
-		const window = Window.createWindow({
+		const mainWindowOptions: BrowserWindowConstructorOptions = {
 			width: 1400,
 			height: 900,
 			frame: false,
 			webPreferences: {
 				preload: path.join(__dirname, "preload")
 			},
-		});
+		};
+
+		const mainWindow = new BrowserWindow(mainWindowOptions);
+
+		mainWindow.removeMenu();
 
 		IPC.init({
-			window,
+			window: mainWindow,
 		});
 
 		if (process.env.ARK_ENABLE_DEV_TOOLS && process.env.ARK_DEV_TOOLS_PATH)
-			await enableDevTools(window, process.env.ARK_DEV_TOOLS_PATH);
+			await enableDevTools(mainWindow, process.env.ARK_DEV_TOOLS_PATH);
 
 		const loadURL =
 			process.env.ARK_ENTRY_URL ||
 			`file://${path.join(__dirname, "../../index.html")}`;
 
-		await window.loadURL(loadURL);
+		await mainWindow.loadURL(loadURL);
 	} catch (e) {
 		console.error(e);
 	}
