+++ electron/tsconfig.json
@@ -11,5 +11,12 @@
 		"isolatedModules": true
 		// "noEmit": false
 	},
-	"include": ["./**/*.ts", "../global.d.ts", "../util/**/*.ts"]
+	"include": [
+		"./**/*.ts",
+		"../global.d.ts",
+		"../util/**/*.ts"
+	],
+	"exclude": [
+		"../node_modules/**"
+	]
 }
