+++ tsconfig.json
@@ -1,7 +1,14 @@
 {
 	"compilerOptions": {
-		"lib": ["dom", "dom.iterable", "esnext"],
-		"types": ["vite/client", "vite-plugin-svgr/client"],
+		"lib": [
+			"dom",
+			"dom.iterable",
+			"esnext"
+		],
+		"types": [
+			"vite/client",
+			"vite-plugin-svgr/client"
+		],
 		"target": "es2020",
 		"module": "ESNext",
 		"strict": true,
@@ -14,6 +21,13 @@
 		"jsx": "react",
 		"noImplicitAny": false
 	},
-	"include": ["src", "./global.d.ts", "util"],
-	"exclude": ["src/mongoshell.d.ts"]
+	"include": [
+		"src",
+		"./global.d.ts",
+		"util"
+	],
+	"exclude": [
+		"src/mongoshell.d.ts",
+		"node_modules/**"
+	]
 }
