--- util/constants.ts
@@ -1,4 +1,0 @@
-export const ERRORS = {
-	AR600: "SSH Tunnel Closed",
-	AR601: "Connection closed",
-};
